+++ web/src/admin/admin-settings/AdminSettingsForm.ts
@@ -14,7 +14,6 @@ import { akFooterLinkInput, IFooterLinkInput } from "./AdminSettingsFooterLinks.
 
 import { DEFAULT_CONFIG } from "#common/api/config";
 
-import { CodeMirrorMode } from "#elements/CodeMirror";
 import { Form } from "#elements/forms/Form";
 
 import { AdminApi, FooterLink, Settings, SettingsRequest } from "@goauthentik/api";
@@ -250,10 +249,7 @@ export class AdminSettingsForm extends Form<SettingsRequest> {
                 help=${msg("Default length of generated tokens")}
             ></ak-number-input>
             <ak-form-element-horizontal label=${msg("Flags")} name="flags" required>
-                <ak-codemirror
-                    mode=${CodeMirrorMode.YAML}
-                    value="${YAML.stringify(settings?.flags ?? {})}"
-                >
+                <ak-codemirror mode="yaml" value="${YAML.stringify(settings?.flags ?? {})}">
                 </ak-codemirror>
                 <p class="pf-c-form__helper-text">
                     ${msg("Modify flags to opt into new authentik behaviours early.")}
