+++ web/src/admin/applications/entitlements/ApplicationEntitlementForm.ts
@@ -5,7 +5,6 @@ import "#elements/forms/SearchSelect/index";
 
 import { DEFAULT_CONFIG } from "#common/api/config";
 
-import { CodeMirrorMode } from "#elements/CodeMirror";
 import { ModelForm } from "#elements/forms/ModelForm";
 
 import { ApplicationEntitlement, CoreApi } from "@goauthentik/api";
@@ -64,7 +63,7 @@ export class ApplicationEntitlementForm extends ModelForm<ApplicationEntitlement
             </ak-form-element-horizontal>
             <ak-form-element-horizontal label=${msg("Attributes")} name="attributes">
                 <ak-codemirror
-                    mode=${CodeMirrorMode.YAML}
+                    mode="yaml"
                     value="${YAML.stringify(this.instance?.attributes ?? {})}"
                 >
                 </ak-codemirror>
