+++ web/src/admin/brands/BrandForm.ts
@@ -13,7 +13,6 @@ import "#components/ak-file-search-input";
 import { DEFAULT_CONFIG } from "#common/api/config";
 import { DefaultBrand } from "#common/ui/config";
 
-import { CodeMirrorMode } from "#elements/CodeMirror";
 import { ModelForm } from "#elements/forms/ModelForm";
 
 import { AKLabel } from "#components/ak-label";
@@ -133,7 +132,7 @@ export class BrandForm extends ModelForm<Brand, string> {
 
                         <ak-codemirror
                             id="branding-custom-css"
-                            mode=${CodeMirrorMode.CSS}
+                            mode="css"
                             value="${this.instance?.brandingCustomCss ??
                             DefaultBrand.brandingCustomCss}"
                         >
@@ -304,7 +303,7 @@ export class BrandForm extends ModelForm<Brand, string> {
                         <ak-codemirror
                             id="attributes"
                             name="attributes"
-                            mode=${CodeMirrorMode.YAML}
+                            mode="yaml"
                             value="${YAML.stringify(this.instance?.attributes ?? {})}"
                             aria-describedby="attributes-help"
                         >
